+++ src/blocks/heading-with-icon/edit.tsx
@@ -1,84 +1,33 @@
 import type { InterpretedAttributes, Attributes } from "./attributes";
 import type { InterpretedUsedContext } from "./context";
 import type { Supports } from "./supports";
-import type {
-	BlockInstanceAsObject,
-	CreateBlockEditProps,
-} from "@atomicsmash/blocks-helpers";
+import type { CreateBlockEditProps } from "@atomicsmash/blocks-helpers";
 import type { CSSProperties } from "react";
 import {
 	InspectorControls,
 	useBlockProps,
 	RichText,
 	BlockControls,
-	store as blockEditorStore,
 } from "@wordpress/block-editor";
-import { store as blocksStore } from "@wordpress/blocks";
 import {
 	Panel,
 	PanelBody,
 	ToolbarGroup,
 	__experimentalUnitControl as UnitControl,
 } from "@wordpress/components";
-import { useSelect } from "@wordpress/data";
-import TokenList from "@wordpress/token-list";
 import { ColourSelectControl, HeadingLevelSelect } from "../helpers.editor";
 import { IconSelectControl, getIcons } from "../svgs.editor";
 import { attributes as definedAttributeOptions } from "./attributes";
-import blockMetaData from "./block.json";
 
 export type BlockEditProps = CreateBlockEditProps<
 	InterpretedAttributes,
 	InterpretedUsedContext
 >;
 
-export function Edit({ clientId, attributes, setAttributes }: BlockEditProps) {
+export function Edit({ attributes, setAttributes }: BlockEditProps) {
 	const { headerElement, headerContent, iconName, library, size, iconColour } =
 		attributes;
 
-	const { thisBlocksStyles } = useSelect((select) => {
-		const { getBlockStyles } = select(blocksStore) as {
-			getBlockStyles: (
-				name: string,
-			) => { name: string; label: string; isDefault?: boolean }[] | undefined;
-		};
-		return {
-			thisBlocksStyles: getBlockStyles(blockMetaData.name),
-		};
-	}, []);
-
-	const { activeStyle } = useSelect(
-		(select) => {
-			const { getBlock } = select(blockEditorStore) as {
-				getBlock: (clientId: string) => BlockInstanceAsObject & {
-					clientId: string;
-					isValid: boolean;
-					originalContent: string;
-					validationIssues: unknown[];
-				};
-			};
-			const className =
-				(getBlock(clientId).attributes?.className as string | undefined) ?? "";
-
-			for (const style of new TokenList(className).values()) {
-				if (!style.includes("is-style-")) {
-					continue;
-				}
-
-				const potentialStyleName = style.substring(9);
-				const activeStyle = thisBlocksStyles?.find(
-					({ name }) => name === potentialStyleName,
-				);
-				if (activeStyle) {
-					return { activeStyle };
-				}
-			}
-			return {
-				activeStyle: thisBlocksStyles?.find((style) => style.isDefault),
-			};
-		},
-		[clientId, thisBlocksStyles],
-	);
 	const blockProps = useBlockProps({
 		style: {
 			"--icon-size": size,
@@ -150,7 +99,6 @@ export function Edit({ clientId, attributes, setAttributes }: BlockEditProps) {
 							headerContent: newHeaderContent,
 						});
 					}}
-					className={activeStyle ? `is-style-${activeStyle.name}` : undefined}
 					placeholder={"Heading"}
 				/>
 			</div>