+++ src/blocks/heading-with-icon/render.php
@@ -31,18 +31,6 @@ if ( isset( $attributes['iconColour'] ) ) {
 	$wrapper_style['--icon-colour'] = $attributes['iconColour'];
 }
 
-$block_classes = isset( $block->parsed_block['attrs']['className'] ) && $block->parsed_block['attrs']['className'];
-
-$active_style_class_name = '';
-if ( $block_classes ) {
-	foreach ( explode( ' ', $block_classes ) as $class_name ) {
-		if ( ! str_starts_with( $class_name, 'is-style' ) ) {
-			continue;
-		}
-		$active_style_class_name = $class_name;
-	}
-}
-
 $icon_renderers = LaunchpadBlocks\Helpers\get_icon_renderers();
 
 
@@ -66,7 +54,7 @@ echo wp_kses_data(
 		}
 		?>
 	</div>
-	<<?php echo esc_html( $attributes['headerElement'] ); ?> class="<?php echo esc_attr( $active_style_class_name ); ?>">
+	<<?php echo esc_html( $attributes['headerElement'] ); ?>>
 	<?php echo wp_kses_post( $attributes['headerContent'] ); ?></<?php echo esc_html( $attributes['headerElement'] ); ?>>
 	<?php echo wp_kses_post( $content ); ?>
 </div>